requirements.txt
Django==3.2.2
django-ckeditor==6.0.0
django-heroku==0.3.1
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
python-slugify==5.0.2
whitenoise==5.2.0